feat: add more features to reduce dependencies (#239)
When importing sub crates directly (like artifacts) we don't need regex,
walkdir, and other deps.

No functional changes intended.
