+++ crates/core/Cargo.toml
@@ -18,25 +18,37 @@ workspace = true
 alloy-primitives.workspace = true
 cfg-if.workspace = true
 dunce.workspace = true
-path-slash.workspace = true
-regex.workspace = true
 semver.workspace = true
 serde_json.workspace = true
 serde.workspace = true
 thiserror.workspace = true
-walkdir.workspace = true
 
+# regex
+regex = { workspace = true, optional = true }
+
+# walkdir
+walkdir = { workspace = true, optional = true }
+
+# svm-solc
 svm = { workspace = true, optional = true }
+# async
 tokio = { workspace = true, optional = true }
 
+# project-util
 tempfile = { workspace = true, optional = true }
 fs_extra = { version = "1.3", optional = true }
+
+[target.'cfg(windows)'.dependencies]
+path-slash.workspace = true
 
 [dev-dependencies]
 tempfile.workspace = true
 
 [features]
-project-util = ["dep:tempfile", "dep:fs_extra"]
-svm-solc = ["dep:svm", "dep:tokio"]
 async = ["dep:tokio"]
+project-util = ["dep:tempfile", "dep:fs_extra"]
+regex = ["dep:regex"]
+svm-solc = ["dep:svm", "dep:tokio"]
+walkdir = ["dep:walkdir"]
+
 test-utils = ["dep:tempfile"]
